Deleted `GPIOCommand` and replaced w/ `IOCommand`
