--- src/action/commands/mod.rs
@@ -1,5 +1,0 @@
-mod gpio;
-mod notifier;
-
-pub use gpio::*;
-pub use notifier::*;
