+++ src/action/mod.rs
@@ -1,6 +1,5 @@
 //! Cybernetic abstractions
 mod command;
-mod commands;
 mod publisher;
 mod routine;
 mod subscriber;
@@ -8,7 +7,6 @@ mod subscribers;
 mod types;
 
 pub use command::*;
-pub use commands::*;
 pub use publisher::*;
 pub use routine::*;
 pub use subscriber::*;
